% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisec_optim.R
\name{base}
\alias{base}
\title{decimal to base-n}
\usage{
base(x, radix, n.digit)
}
\arguments{
\item{x}{integer scalar. The number to transform}

\item{radix}{integer scalar. radix of the target number, \eqn{2} for binary,
\eqn{8} for octal, etc.}

\item{n.digit}{integer scalar. how many digits the result contains, unused
uppper digits are filled by \eqn{0}}
}
\value{
integer. each scalar of the vector represent one digit of the
transformed number
}
\description{
transform decimal integer to other number mode, such as binary
(\code{radix = 2})
}
\examples{
base(443L, 3L, 6L)
}
