% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisec-optim.R
\name{biosec_optim_impl}
\alias{biosec_optim_impl}
\title{working horse for \link{biosec_optim}}
\usage{
biosec_optim_impl(fun, spaces, partition, trim, enlarge)
}
\arguments{
\item{spaces}{numeric array. \code{spaces[ , , k]} should
give a \code{space}, see \code{bisec_optim} below}
}
\value{
list. list of \code{list(fun.val, subsapce)} , first is best
}
