% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisec-fit.R
\name{bisec_fit}
\alias{bisec_fit}
\title{General-purpose fit}
\usage{
bisec_fit(data, ff, extra, space, partition, times, trim, enlarge)
}
\arguments{
\item{data}{numeric matrix. \code{data[i, 1]} gives \eqn{xi} and \code{data[i, 2]} gives
\eqn{yi}}

\item{ff}{function. \code{ff(para, extra)} gives the function \code{f()} that
\eqn{y=f(x)} where the range of para make a \code{space}.}

\item{extra}{numeric vector. see \code{ff}}

\item{space, partition, times, trim, enlarge}{see \link{bisec_optim}}
}
\value{
list of list.
\enumerate{
\item information about subrange1
\enumerate{
\item \eqn{R^2} using median of subrange1 as \code{para} for \code{ff}
\item subrange1
}
\item information about subrange2
\enumerate{
\item \eqn{R^2} using median of subrange2 as \code{para} for \code{ff}
\item subrange1
}
\item ...
}
}
\description{
you can almost fit any function you want
}
\details{
even complex coupled differential equation that you can't get a analytic solution is ok.
}
\section{to do}{

1. support multivariate function
}

\examples{
bisec_fit(data1, f_protein, extra11, space1, 2, 5, 1/3, 1)
bisec_fit(data2, f_AI2_out, extra2, space2, 2, 5, 4, 1)

}
\seealso{
\link{biosec_optim}
}
