% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI2-case.R
\name{f_AI2_out}
\alias{f_AI2_out}
\title{function for extracelluar AI2 concentration}
\usage{
f_AI2_out(para, extra)
}
\arguments{
\item{para}{numeric. unknown parameters}

\item{extra}{numeric, given parameters}
}
\value{
function. returns \eqn{y} when giving \eqn{x} as argument
}
\description{
function for extracelluar AI2 concentration
}
\examples{
NULL

}
\seealso{
Other AI-2 case functions: 
\code{\link{f_protein}()}
}
\concept{AI-2 case functions}
