% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI2-case.R
\name{f_protein}
\alias{f_protein}
\title{function for protein concentration}
\usage{
f_protein(para, extra)
}
\arguments{
\item{para}{numeric. unknown parameters}

\item{extra}{numeric, given parameters}
}
\value{
function. returns \eqn{y} when giving \eqn{x} as argument
}
\description{
function for protein concentration
}
\examples{
NULL

}
\seealso{
Other AI-2 case functions: \code{\link{f_AI2_out}}
}
\concept{AI-2 case functions}
