% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ff_para_extra}
\alias{ff_para_extra}
\title{Title}
\usage{
ff_para_extra()
}
\arguments{
\item{ff}{function. `ff(para, extra)` gives the function `f()` that
\eqn{y=f(x)}}

\item{para}{numeric. see \code{ff}}

\item{extra}{numeric. see \code{ff}}
}
\description{
Title
}
