% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{R_square}
\alias{R_square}
\title{Title}
\usage{
R_square(data, ff, para, extra)
}
\arguments{
\item{data}{numeric matrix. \code{data[,1]} gives \eqn{x} and \code{data[,2]} gives
\eqn{y}}

\item{ff}{function. \code{ff(para, extra)} gives the function \code{f()} that
\eqn{y=f(x)}}

\item{para}{numeric. see \code{ff}}

\item{extra}{numeric. see \code{ff}}
}
\value{
numeric scalar. \eqn{R^2} (coefficient of determination) showing how
well the function fits \code{data}
}
\description{
Title
}
\examples{
\\donotrun {
    R_square(data1, f_protein, ideal.para11, extra11)
    R_square(data1, f_protein, ideal.para11, extra12)
    R_square(data2, f_AI2_out, ideal.para2, extra2)
}
}
