% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-fit.R
\name{simulate_data}
\alias{simulate_data}
\title{generate simulated data from given model}
\usage{
simulate_data(ff, para, extra, x)
}
\arguments{
\item{x}{numeric. used to generate \code{y} from \verb{ff(para, extra)(x))}}
}
\value{
numeric matrix. \code{cbind(x, y)}
}
\description{
generate simulated data from given model
}
\examples{
simulate_data(f_protein, ideal.para11, extra11, seq(60,240,60))
simulate_data(f_AI2_out, ideal.para2, extra2, seq(0,270,30))

}
\seealso{
Other simulate functions: 
\code{\link{simulate_partition_fit}()}
}
\concept{simulate functions}
