% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-fit.R
\name{simulate_partition_fit}
\alias{simulate_partition_fit}
\title{use simulated data to evaluate fitting function}
\usage{
simulate_partition_fit(x, ff, ideal.para, extra, patition, times, trim,
  enlarge, scale, details = TRUE)
}
\arguments{
\item{x}{numeric. see \link{simulate_data}}

\item{ff}{function. see \link{simulate_data}}

\item{ideal.para}{numeric. see \link{simulate_data}. ideal value for \code{para}}

\item{times}{integer scalar. see \link{partition_fit}}

\item{scale}{integer scalar. how many times is \code{ideal.para} enlarged to
generate the initial range}

\item{details}{logical scalar. whether result should include extra
information.}
}
\value{
vector. see examples for details.
}
\description{
use simulated data to evaluate fitting function
}
\examples{
simulate_partition_fit(seq(60,240,60), f_protein, ideal.para11 , extra11, 10,3,1/3, 1, 3)
simulate_partition_fit(seq(0,270,30), f_AI2_out, ideal.para2 , extra2, 3,1,1/3, 1, 3)

}
\seealso{
Other simulate functions: \code{\link{simulate_data}}
}
\concept{simulate functions}
