% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisec-fit.R
\name{assess}
\alias{assess}
\title{assess how reality accord with the ideal}
\usage{
assess(reality, ideal)
}
\arguments{
\item{reality}{numeric. real parameters}

\item{ideal}{numeric. ideal parameters}
}
\value{
numeric scalar. the relative similarity between \code{reality} and \code{ideal}
}
\description{
assess how reality accord with the ideal
}
\examples{
assess(ideal.para11, ideal.para11 * (1 + 0.1 * runif(3)))

}
