% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisec-optim.R
\name{bisec_optim}
\alias{bisec_optim}
\title{General-purpose Optimization (Minimization)}
\usage{
bisec_optim(fun, space, partition, times, trim, enlarge)
}
\arguments{
\item{fun}{function. the function to be optimized, i.e. find minimal value. \code{fun(c(para1, para2, ..., paran))} should return a numeric scalar where \code{para} is a numeric vector, such as}

\item{space}{numeric matrix. \eqn{n*2} matrix, where \eqn{n} is the number of
parameters to fit. \code{space[i, ]} should give a the range of ith parameter, i.e. \code{c(para1.min, para2.max)}. \code{space} specifies a n-dimension space, this space is uniformly partitioned into subspaces, and those subspaces which have the highest score is reserved for further partitioning. Finally we get the best space, and the centre point of this space contains the parameters we want.}

\item{partition}{integer scalar. how partitions are the original range
divided per parameter.}

\item{times}{integer scalar. how many times should the initial range be partitioned.}

\item{trim}{numeric scalar. must be positive, determine what proportion of total subranges will
be reserved for the next round of partition. Let \code{n} be the number of subspaces per space, if \code{trim} >= 1, \code{as.integer(trim)}; otherwise, \code{ceiling(n^trim)}.}

\item{enlarge}{integer scalar. how many times is the selected subrange
enlarged before the next round of partition.}
}
\value{
list.
\enumerate{
\item value, numeric saclar. smallest function value found
\item para, numeric vector. best parameters.
\item space, numeric matrix. best solution space.
}
}
\description{
\code{bisec_optim()} found parameters which minimize target function in a given space
}
\details{
see paras.
}
\section{to do}{

1. partition can be a integer vector, i.e. different parameter can be partitioned for different times
}

\examples{
bisec_optim(sum, matrix(c(0,1,0,1,0,1), 3, byrow = T), 3, 10, 4, 1)
bisec_optim(sum, matrix(c(0,1,0,1,0,1), 3, byrow = T), 3, 10, 1/3, 1)

}
\seealso{
\link{biosec_optim_impl}
}
