% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI2-case.R
\name{f_AI2_out}
\alias{f_AI2_out}
\title{Title}
\usage{
f_AI2_out(para, extra)
}
\arguments{
\item{para}{numeric. unknown parameters}

\item{extra}{numeric, given parameters}
}
\value{
function. returns \eqn{y} when giving \eqn{x} as argument
}
\description{
Title
}
\examples{
f_AI2_out(ideal.para2, extra2)(seq(0, 270, 30))
}
\seealso{
Other AI-2 case functions: \code{\link{f_protein}}
}
