% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI2-case.R
\name{f_protein}
\alias{f_protein}
\title{Title}
\usage{
f_protein(para, extra)
}
\arguments{
\item{para}{numeric. unknown parameters}

\item{extra}{numeric, given parameters}
}
\value{
function. returns \eqn{y} when giving \eqn{x} as argument
}
\description{
Title
}
\examples{
donotrun {
    protein_ACDB <- f_protein(ideal.para11, extra11) #protein with ACDB gene and IPTG
    protein_K <- f_protein(ideal.para11, extra12) #protein with IPTG
    protein <- f_protein(ideal.para12, extra12) #protein with nothing
}
}
\seealso{
Other AI-2 case functions: \code{\link{f_AI2_out}}
}
\concept{AI-2 case functions}
