% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisec-fit.R
\name{R_square}
\alias{R_square}
\title{get R square from data and fit function}
\usage{
R_square(data, ff, para, extra)
}
\arguments{
\item{data}{numeric matrix. \code{data[,1]} gives \eqn{x} and \code{data[,2]} gives
\eqn{y}}
}
\value{
numeric scalar. \eqn{R^2} (coefficient of determination) showing how
well the function fits \code{data}
}
\description{
get R square from data and fit function
}
\examples{
R_square(data1, f_protein, ideal.para11, extra11)
R_square(data1, f_protein, ideal.para11, extra12)
R_square(data2, f_AI2_out, ideal.para2, extra2)

}
