% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-fit.R
\name{simulate_data}
\alias{simulate_data}
\title{Title}
\usage{
simulate_data(ff, para, extra, x)
}
\arguments{
\item{x}{numeric. used to generate \code{y} from \code{ff(para, extra)(x))}}
}
\value{
numeric matrix. \code{cbind(x, y)}
}
\description{
Title
}
\examples{
\\donotrun {
    data1 <- simulate_data(f_protein, ideal.para11, extra11, seq(60,240,60));
    data2 <- simulate_data(f_AI2_out, ideal.para2, extra21, seq(0,270,30));
}
}
\seealso{
Other simulate functions: \code{\link{simulate_partition_fit}}
}
\concept{simulate functions}
